fix: move sources to src/ folter, refactor _sumPrivkeys
