+++ tsconfig.json
@@ -1,4 +1,10 @@
 {
   "extends": "@tsconfig/recommended/tsconfig.json",
-  "include": ["index.ts", "noble_ecc.ts"]
+  "compilerOptions": {
+    "outDir": "./dist",
+    "declaration": true,
+  },
+  "include": [
+    "src/**/*"
+  ]
 }
